refactor(web-server): fix cross-route dep + extract task-usage sub-zone
Two architectural improvements to address web-server zone's catastrophic
cohesion/coupling metrics (0.20/0.80):

1. **Fix routes-hench → routes-status cross-route import**
   routes-hench.ts imported clearStatusCache() directly from routes-status.ts
   (a peer route module), creating a horizontal coupling between sibling routes.
   Fixed via injection: handleHenchRoute now accepts HenchRouteOptions with
   onStatusInvalidate?, injected by start.ts. Documents the seam in CLAUDE.md.

2. **Extract task-usage analytics to task-usage/ sub-zone**
   Moved incremental-task-usage.ts, usage-cleanup-scheduler.ts, and
   register-scheduler.ts from server/ flat directory into task-usage/ subdirectory.
   The existing task-usage.ts facade already grouped these logically; the
   subdirectory makes the zone boundary visible in the file tree and enables
   SourceVision's community detection to identify them as a distinct zone.
   Old paths become thin re-export shims for backward compatibility.
   Updated test imports and build-output-contract to canonical paths.
